Create Challenge: Cycling race
Ava is participating in a cycling race. The race route lies along the shore of a lake and exclusively follows its edges. The lake's shore consists only of straight vertical or horizontal segments, and each segment moves in one of the four cardinal directions: north, south, east, or west.
We introduce a coordinate system where the OX-axis represents west-to-east and the OY-axis represents south-to-north. The starting point of the race is the first point provided, and the route ends back at this starting point. During each straight segment, participants move in one of the four cardinal directions, and they only change direction at corners. The participants never reverse along the same segment, meaning they don't change direction from north to south or from east to west (or vice versa).
Ava is young and occasionally gets nervous at certain corners. If Ava fails to maneuver properly at a corner and slips into the water, it will be very unpleasant for her. In other words, Ava considers such corners, where a slip would immediately lead to the water, "dangerous."
Help Ava prepare for the race. Write a program to determine the number of dangerous corners along the race route.
Note: At the starting point, there is a stop, so this point should not be considered a dangerous corner.
Input
The first line of input contains an even integer n representing the number of straight segments in the route.
The next n lines each contain two integers x_i, y_i representing the coordinates of the i-th corner along the route. The first point is the starting point of the race, and the i-th point represents the coordinates of the i-th corner.
4≤n≤2004
 −10000≤xi, yi≤10000
It is guaranteed that:
The race route does not intersect itself except at corners.
The direction always changes at each corner, and all segments move in one of the four cardinal directions.
Output
Print the number of dangerous corners.
Example
Sample Input 1
6  
0 0  
0 1  
1 1  
1 2  
2 2  
2 0
Sample Output 1
1
The race route for this example is shown below, where the point (1,1)(1,1) is a dangerous corner:
Sample Input 2
6  
1 1  
1 2  
2 2  
2 0  
0 0  
0 1
Sample Output 2
0
Sample Input 3
4 
0 0  
0 -5  
10 -5  
10 0 
Sample Output 3
0